requirements.txt
python-dotenv
cryptography
gspread-pandas
pyyaml